* commit 'faa14fc1f215adf405a36b6f817c0559ed109ad1':
  fix test
  fix paths
  fix paths
  test python only for windows release
  python for windows
  fix windows warnings
  return coverage tests
  fix tests
  use TCP_NO_DELAY for small messages performance, test for getlast
  fix win?
  fix win?
  test json
  python api for linux only
  fix test
  fix,temporarily switch off coverage
  do not save beamline/beamtime in the file name
  add numpy include dirs
  removed python calls from c
  more work on python API
  start working at python API for

+++ producer/api/src/request_handler_tcp.cpp
@@ -55,7 +55,7 @@ Error RequestHandlerTcp::SendHeaderAndData(const ProducerRequest* request) {
         return io_error;
     }
 
-    io__->Send(sd_, (void*) request->data.get(), request->header.data_size, &io_error);
+    io__->Send(sd_, (void*) request->data.get(), (size_t)request->header.data_size, &io_error);
     if(io_error) {
         return io_error;
     }
